Add .start and .stop methods for Reporter
Allows executing the block of code to be profiled to be executed
independently of the starting and stopping of the object tracing.
Useful if you don't have direct access to the block to be profiled, and
only access wrapper methods around the block.

Also updated the report to be stored as an instance variable for later
access, instead of dumping it directly from the `.stop` call.

`.run` now also uses these interfaces since it is the same that is used
under the hood by `ObjectSpace.trace_object_allocations`, and `.run` and
`.report` now become a quick interface for running the:

    `.new` -> `.start` -> `&block.call` -> `.stop`

workflow in the codebase.