Move all ::report tests to use create_report
There were a few tests in TestReport that didn't flow through
create_report, but instead used it's own MemoryProfiler::Reporter.report
block to capture the results.

Changing report_block to accept a block, and having create_report pass
that block on to report_block allows the same method interface to be
used accross all of the tests, regardless of the API being tested in
create_report.