Use inheritance for testing multiple report APIs
By using TestReporter as the base class, we now can inherit from that
class to then test other forms of the Reporter interface (.start and
.stop and MemoryProfiler::start and MemoryProfiler::stop) and only need
to overwrite the `create_report` method in the inherited class.  Results
should remain the same, and only how the results are gathered should
change.

Any specific tests to that interface can then live within that test file
itself, making it clearer where edge cases exists in different APIs of
the MemoryProfiler::Reporter.

Note:  This commit is basically moving code around, and the only
additions to the amount of tests being run is because the subclasses are
now inheriting from TestReporter, which now allows the sub classes to
take advantage of the `create_report` changes from the previous commit,
and weren't being tested in the same copacity prior.