Add md5 to MemoryProfiler::Stat
Future option to be able to compare string values using a md5 value
instead of the full string.

Also, adds a test around the output of the "String Report", confirming
that it works with the current implementation.  Implementation of the
md5 into the String Report will be introduced in a future commit.