Add :mem_lite option for Reporter
Allows the MemoryProfiler::Reporter to be run in less memory intensive
way, at the cost of some speed in generating the report.  This is ideal
in cases where there are a lot of objects that can be expanded into
strings, and will be duplicated/create in the process generating the
report while iterating over the ObjectSpace.

The `hexdigest`'ing is offset a bit by the cache (when in scenarios
where that is helpful), but in the end, the bottle neck is the digests.

This is an opt in setting, and the previous implementation is still
unaffected and the default.

Moves gathering the `memsize` variable to before the string copying
portion of the MemoryProfiler::Reporter#object_list method.  This was
causing issues when using `:memlite`, and the values were not equal when
compared against using without.  This should not affect the reported
result values that were gathered previously.